requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
blueprint==3.4.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.49.4
MarkupSafe==3.0.1
multidict==6.1.0
openai==1.51.2
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
setuptools==69.0.3
smmap==5.0.1
sniffio==1.3.1
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.15.3
zipp==3.20.2
google-cloud-storage==2.18.2
google-cloud-container==2.53.0
google-auth==2.33.0
docker==6.1.3
